+++ test/clj/rems/application/test_model.clj
@@ -1,7 +1,7 @@
 (ns rems.application.test-model
   (:require [clojure.set :as set]
             [clojure.string :as str]
-            [clojure.test :refer :all]
+            [clojure.test :refer [deftest is testing use-fixtures]]
             [hiccup.core :as hiccup]
             [medley.core :refer [map-vals]]
             [rems.api.schema :as schema]
@@ -490,7 +490,6 @@
 
 (deftest test-application-view-saved
   (is (= saved-application (recreate saved-application))))
-
 
 (def licenses-accepted-event {:event/type :application.event/licenses-accepted
                               :event/time (DateTime. 2500)
@@ -1254,7 +1253,8 @@
                                 :application/forms [{:form/id 40}]
                                 :workflow/id 50
                                 :workflow/type :workflow/master
-                                :application/licenses [{:license/id 30} {:license/id 31} {:license/id 32}]}
+                                :application/licenses [{:license/id 30} {:license/id 31} {:license/id 32}]
+                                :event/visibility :visibility/public}
                                {:event/type :application.event/draft-saved
                                 :application/id 1
                                 :event/time (DateTime. 2000)
@@ -1263,24 +1263,28 @@
                                                            {:form 40 :field "42" :value "bar"}
                                                            {:form 40 :field "43" :value "private answer"}
                                                            {:form 40 :field "field-does-not-exist" :value "something"}]
-                                :event/actor-attributes {:userid "applicant" :email "applicant@example.com" :name "Applicant" :secret "secret"}}
+                                :event/actor-attributes {:userid "applicant" :email "applicant@example.com" :name "Applicant" :secret "secret"}
+                                :event/visibility :visibility/public}
                                {:event/type :application.event/licenses-accepted
                                 :application/id 1
                                 :event/time (DateTime. 2500)
                                 :event/actor "applicant"
                                 :application/accepted-licenses #{30 31 32}
-                                :event/actor-attributes {:userid "applicant" :email "applicant@example.com" :name "Applicant" :secret "secret"}}
+                                :event/actor-attributes {:userid "applicant" :email "applicant@example.com" :name "Applicant" :secret "secret"}
+                                :event/visibility :visibility/public}
                                {:event/type :application.event/submitted
                                 :application/id 1
                                 :event/time (DateTime. 3000)
                                 :event/actor "applicant"
-                                :event/actor-attributes {:userid "applicant" :email "applicant@example.com" :name "Applicant" :secret "secret"}}
+                                :event/actor-attributes {:userid "applicant" :email "applicant@example.com" :name "Applicant" :secret "secret"}
+                                :event/visibility :visibility/public}
                                {:event/type :application.event/approved
                                 :application/id 1
                                 :event/time (DateTime. 4000)
                                 :event/actor "handler"
                                 :application/comment "looks good"
-                                :event/actor-attributes {:userid "handler" :email "handler@example.com" :name "Handler" :secret "secret"}}]
+                                :event/actor-attributes {:userid "handler" :email "handler@example.com" :name "Handler" :secret "secret"}
+                                :event/visibility :visibility/public}]
           :application/user-roles {"handler" #{:handler}, "reporter1" #{:reporter}}
           :application/role-permissions nil
           :application/description "foo"
@@ -1331,6 +1335,7 @@
             :event/time (DateTime. 1)
             :event/actor "applicant"
             :event/actor-attributes {:userid "applicant" :email "applicant@example.com" :name "Applicant" :secret "secret"}
+            :event/visibility :visibility/public
             :application/id 1
             :application/resources [{:catalogue-item/id 10
                                      :resource/id 11
@@ -1366,6 +1371,7 @@
             :event/time (DateTime. 1)
             :event/actor "handler"
             :event/actor-attributes {:userid "handler" :email "handler@example.com" :name "Handler" :secret "secret"}
+            :event/visibility :visibility/handling-users
             :application/id 1
             :application/deciders [{:userid "decider" :email "decider@example.com" :name "Decider" :secret "secret"}
                                    {:userid "reviewer" :email "reviewer@example.com" :name "Reviewer" :secret "secret"}]}
@@ -1380,6 +1386,7 @@
             :event/time (DateTime. 1)
             :event/actor "handler"
             :event/actor-attributes {:userid "handler" :email "handler@example.com" :name "Handler" :secret "secret"}
+            :event/visibility :visibility/handling-users
             :application/id 1
             :application/reviewers [{:userid "decider" :email "decider@example.com" :name "Decider" :secret "secret"}
                                     {:userid "reviewer" :email "reviewer@example.com" :name "Reviewer" :secret "secret"}]}
@@ -1394,6 +1401,7 @@
             :event/time (DateTime. 1)
             :event/actor "handler"
             :event/actor-attributes {:userid "handler" :email "handler@example.com" :name "Handler" :secret "secret"}
+            :event/visibility :visibility/public
             :application/id 1
             :application/member {:userid "member" :email "member@example.com" :name "Member" :secret "secret"}}
            (model/enrich-event {:event/type :application.event/member-added
@@ -1407,6 +1415,7 @@
             :event/time (DateTime. 1)
             :event/actor "handler"
             :event/actor-attributes {:userid "handler" :email "handler@example.com" :name "Handler" :secret "secret"}
+            :event/visibility :visibility/public
             :application/id 1
             :application/member {:userid "member" :email "member@example.com" :name "Member" :secret "secret"}}
            (model/enrich-event {:event/type :application.event/member-removed
@@ -1635,11 +1644,11 @@
                           (model/application-view {:event/type :application.event/remarked
                                                    :application/comment "this is public"
                                                    :event/actor "bob"
-                                                   :application/public true})
+                                                   :event/public true})
                           (model/application-view {:event/type :application.event/remarked
                                                    :application/comment "this is private"
                                                    :event/actor "smith"
-                                                   :application/public false}))
+                                                   :event/public false}))
           enriched (-> application
                        (permissions/update-role-permissions {:role-1 #{:see-everything}})
                        (model/enrich-with-injections injections))]